+++ src/checkbox.jsx
@@ -23,7 +23,8 @@ module.exports = React.createClass({
         onKeyDown             : React.PropTypes.func,
         onKeyUp               : React.PropTypes.func,
         onKeyPress            : React.PropTypes.func,
-        className             : React.PropTypes.string
+        className             : React.PropTypes.string,
+        tabIndex              : React.PropTypes.string
     },
 
     getDefaultProps()
@@ -108,6 +109,7 @@ module.exports = React.createClass({
                 onKeyDown   = {this.onKeyDown}
                 onKeyPress  = {this.onKeyPress}
                 disabled    = {this.props.disabled}
+                tabIndex    = {this.props.tabIndex}
             />
         );
     }